+++ spear/utils/data_editor.py
@@ -69,7 +69,7 @@ def get_data(path, check_shapes = True, class_map = None):
 			assert type(data[i]) == np.ndarray
 		data.append(pickle.load(file))
 	
-	assert type(data[9]) == np.int
+	assert type(data[9]) == int
 
 	if check_shapes:
 		assert data[1].shape == data[2].shape # l, m
@@ -106,7 +106,6 @@ def get_data(path, check_shapes = True, class_map = None):
 			data[i] = data[i].astype(np.float32)
 		else:
 			data[i] = data[i].astype(np.int32)
-
 	return data
 
 def get_classes(path):
